+++ src/song-data.js
@@ -1,0 +1,319 @@
+export default [
+  {
+    lyrics: {
+      0: "oli",
+      1: "tunnelma",
+      2: "niin",
+      3: "huumaava"
+    },
+    artist: "Rauli Badding Somerjoki",
+    song: "Paratiisi"
+  },
+  {
+    lyrics: {
+      0: "juokse",
+      1: "sinne",
+      2: "missä",
+      3: "on",
+      4: "viileää"
+    },
+    artist: "Kaija Koo",
+    song: "Tinakenkätyttö"
+  },
+  {
+    lyrics: {
+      0: "hei",
+      1: "me",
+      2: "juhliin",
+      3: "lähdetään",
+      4: "et",
+      5: "pääse"
+    },
+    artist: "Nylon Beat",
+    song: "Viimeinen"
+  },
+  {
+    lyrics: {
+      0: "💯",
+      1: "⚡️",
+      2: "👊",
+      3: "🔥"
+    },
+    artist: "Vicky Rosti / Antti Tuisku",
+    song: "Sata Salamaa"
+  },
+  {
+    lyrics: {
+      0: "all",
+      1: "we",
+      2: "need",
+      3: "is",
+      4: "lightning",
+      5: "with",
+      6: "power",
+      7: "and",
+      8: "might"
+    },
+    artist: "Lordi",
+    song: "Hard Rock Hallelujah"
+  },
+  {
+    lyrics: {
+      0: "sinulle",
+      1: "mä",
+      2: "antaisin",
+      3: "sydämen",
+      4: "jos",
+      5: "sä"
+    },
+    artist: "Jari Sillanpää",
+    song: "Satulinna"
+  },
+  {
+    lyrics: {
+      0: "samat",
+      1: "tuulet",
+      2: "puhaltaa",
+      3: "samat",
+      4: "sateet",
+      5: "lankeaa"
+    },
+    artist: "Anssi Kela",
+    song: "Puistossa"
+  },
+  {
+    lyrics: {
+      0: "with",
+      1: "a taste",
+      2: "of",
+      3: "your",
+      4: "lips",
+      5: "i'm",
+      6: "on"
+    },
+    artist: "Britney Spears",
+    song: "Toxic"
+  },
+  {
+    lyrics: {
+      0: "rakkaus",
+      1: "on",
+      2: "kuollut",
+      3: "elämän",
+      4: "virtaan"
+    },
+    artist: "Jonna Tervomaa",
+    song: "Rakkauden haudalla"
+  },
+  {
+    lyrics: {
+      0: "läpi",
+      1: "yön",
+      2: "se",
+      3: "vie",
+      4: "ja",
+      5: "toivon",
+      6: "saat"
+    },
+    artist: "Leijonakuningas",
+    song: "Tie elämään"
+  },
+  {
+    lyrics: {
+      0: "tahtoisin",
+      1: "oluen",
+      2: "mut",
+      3: "ei",
+      4: "en",
+      5: "saa"
+    },
+    artist: "Kasmir",
+    song: "Vadelmavene"
+  },
+  {
+    lyrics: {
+      0: "en",
+      1: "oo",
+      2: "enää",
+      3: "sinisilmäinen",
+      4: "oon"
+    },
+    artist: "Frederik",
+    song: "Kolmekymppinen"
+  },
+  {
+    lyrics: {
+      0: "syvälle",
+      1: "se",
+      2: "upposi",
+      3: "niin",
+      4: "aivan",
+      5: "kuin",
+    },
+    artist: "Siiri Nordin",
+    song: "Sydämeni osuman sai"
+  },
+  {
+    lyrics: {
+      0: "usko",
+      1: "pois",
+      2: "että",
+      3: "onnellinen",
+      4: "olen",
+      5: "vaan"
+    },
+    artist: "Popeda",
+    song: "Lihaa ja perunaa"
+  },
+  {
+    lyrics: {
+      0: "ja",
+      1: "mietin",
+      2: "mikä",
+      3: "tunne",
+      4: "tää",
+      5: "on",
+    },
+    artist: "Haloo Helsinki",
+    song: "Beibi"
+  },
+  {
+    lyrics: {
+      0: "oon",
+      1: "vain",
+      2: "sulle",
+      3: "ihminen",
+      4: "villi"
+    },
+    artist: "Pocahontas",
+    song: "Tuulen värit"
+  },
+  {
+    lyrics: {
+      0: "🍺",
+      1: "🖥",
+      2: "➕",
+      3: "🤠️🦃"
+    },
+    artist: "Pate Mustajärvi",
+    song: "Ukkometso"
+  },
+  {
+    lyrics: {
+      0: "ne",
+      1: "levottomat",
+      2: "illat",
+      3: "sekä",
+      4: "unettomat",
+      5: "yöt"
+    },
+    artist: "Zen Cafe",
+    song: "Todella Kaunis"
+  },
+  {
+    lyrics: {
+      0: "ajetaan",
+      1: "kovaa",
+      2: "ollaan",
+      3: "ihan",
+      4: "hiljaa"
+    },
+    artist: "Vesala",
+    song: "Tequila"
+  },
+  {
+    lyrics: {
+      0: "once",
+      1: "more",
+      2: "you",
+      3: "open",
+      4: "the",
+      5: "door"
+    },
+    artist: "Celine Dion",
+    song: "My Heart Will Go On"
+  },
+  {
+    lyrics: {
+      0: "mä",
+      1: "ensin",
+      2: "näin",
+      3: "vain",
+      4: "meren",
+      5: "sinisen"
+    },
+    artist: "Taiska",
+    song: "Mombasa"
+  },
+  {
+    lyrics: {
+      0: "juon",
+      1: "kaljaa",
+      2: "auringon",
+      3: "nousuun"
+    },
+    artist: "Leevi and the Leavings",
+    song: "Pohjois-Karjala"
+  },
+  {
+    lyrics: {
+      0: "paljain",
+      1: "jaloin",
+      2: "kierrän",
+      3: "maailmaa"
+    },
+    artist: "Tapio Rautavaara",
+    song: "Kulkuri ja joutsen"
+  },
+  {
+    lyrics: {
+      0: "kotiin",
+      1: "suksi",
+      2: "palmun",
+      3: "alta",
+      4: "yhä",
+      5: "näytit"
+    },
+    artist: "Freeman ",
+    song: "Ajetaan tandemilla"
+  },
+  {
+    lyrics: {
+      0: "do",
+      1: "you",
+      2: "feel",
+      3: "my",
+      4: "heart",
+      5: "beating"
+    },
+    artist: "The Bangles",
+    song: "Eternal flame"
+  },
+  {
+    lyrics: {
+      0: "tähti",
+      1: "kirkkain",
+      2: "sua",
+      3: "taas",
+      4: "mä",
+      5: "pyynnöin"
+    },
+    artist: "Katri Helena",
+    song: "Katson sineen taivaan"
+  }
+  /*
+  ,
+  {
+    lyrics: {
+      0: "",
+      1: "",
+      2: "",
+      3: "",
+      4: "",
+      5: ""
+    },
+    artist: "",
+    song: ""
+  }
+  */
+]
